pyproject.toml
[project]
name = "mmm-system"
version = "0.1.0"
description = "Marketing Mix Modeling system with PyMC-Marketing and Google Meridian"
requires-python = ">=3.11"
dependencies = [
    # Data pipeline
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    # Storage
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "duckdb>=0.10.0",
    # Models
    "pymc-marketing>=0.9.0",
    "arviz>=0.18.0",
    # API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    # Orchestration
    "prefect>=2.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
]
meridian = [
    "google-meridian>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
